requirements.txt
# Core LangChain
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13

# LangGraph
langgraph==0.0.20

# Document Processing
pypdf==3.17.4
docx2txt==0.8
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1  # For database migrations

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
